requirements.txt
# Core (required)
pydantic>=2.0

# Optional (enhanced functionality)
pyyaml>=6.0
jinja2>=3.0
watchdog>=3.0

# Development
pytest>=7.0
